Player can hold breath for set amount of time

+++ Assets/07. Scripts/Player/Health.cs
@@ -16,7 +16,7 @@ public class Health : MonoBehaviour
         SetHealthBar();
     }
 
-    public void ChangeHealth(int change)
+    public void ChangeHealth(float change)
     {
         health += change;
         if(health <= 0)
